Kleine Zwischenphase

+++ JLinesV2/Game.swift
@@ -374,9 +374,9 @@ class Game: UIView, Printable {
         constraintsArray.append(NSLayoutConstraint(item: joyStick, attribute: .Bottom, relatedBy: .Equal, toItem: self, attribute: .Bottom, multiplier: 1.0,
             constant:  -GV.joyStickRadius / 2))
         
-        constraintsArray.append(NSLayoutConstraint(item: joyStick, attribute: .Width, relatedBy: .Equal, toItem: nil, attribute: .NotAnAttribute, multiplier: 1.0, constant: GV.joyStickRadius * 1.2))
+        constraintsArray.append(NSLayoutConstraint(item: joyStick, attribute: .Width, relatedBy: .Equal, toItem: nil, attribute: .NotAnAttribute, multiplier: 1.0, constant: CGFloat(Int(GV.joyStickRadius * 1.2))))
         
-        constraintsArray.append(NSLayoutConstraint(item: joyStick, attribute: .Height , relatedBy: .Equal, toItem: nil, attribute: .NotAnAttribute, multiplier: 1.0, constant: GV.joyStickRadius * 1.2))
+        constraintsArray.append(NSLayoutConstraint(item: joyStick, attribute: .Height , relatedBy: .Equal, toItem: nil, attribute: .NotAnAttribute, multiplier: 1.0, constant: CGFloat(Int(GV.joyStickRadius * 1.2))))
         self.addConstraints(constraintsArray)
     }
